Upgrade from OpenGL 2.0 to 3.3
